Code
In the initial section of the code, the clustering method is defined, with the user specifying the number of clusters as 5 based on the input graph visualization. However, if the number of clusters in a given file is unknown, the subsequent part of the code can be employed to determine it. The bar graph generated by this section will display various numbers of clusters, and the one with the highest value indicates the most suitable number of clusters.